requirements.txt
scipy==1.6.2
albumentations==0.5.2
ujson==4.0.1
torch==1.7.0
torchvision==0.8.1
~pencv_python==4.4.0.46
numpy==1.19.2
Pillow==8.3.2
rfconv==0.0.2b20210406
scikit_learn==0.24.2
